some protection are done. Had to check all code one more time to see if everything is fine. Part of norm is done too. Only two functions more than 25 lines.
