+++ srcs/parsing/game_infos.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   game_infos.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tpauvret <tpauvret@student.42lausanne.ch>  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/02 15:44:15 by tpauvret          #+#    #+#             */
+/*   Updated: 2022/05/02 15:55:41 by tpauvret         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../../inc/cub3d.h"
 
 int	check_all_datas(t_game *game)
@@ -30,25 +42,29 @@ int	first_wall_line(char *line)
 	return (0);
 }
 
-int	collect_data(char *line, t_game *game)
+int	collect_data(char *l, t_game *game)
 {
 	int	i;
 
 	i = 0;
-	while (line[i] != 32)
+	if (!l)
+		return (0);
+	while (l[i] != 32)
 		i++;
-	if (!ft_strncmp(line, "NO", i) && game->texture_pack->no->path == NULL)
-		game->texture_pack->no->path = ft_substr(line, i + 1, ft_strlen(line) - i);
-	else if (!ft_strncmp(line, "SO", i) && game->texture_pack->so->path == NULL)
-		game->texture_pack->so->path = ft_substr(line, i + 1, ft_strlen(line) - i);
-	else if (!ft_strncmp(line, "EA", i) && game->texture_pack->ea->path == NULL)
-		game->texture_pack->ea->path = ft_substr(line, i + 1, ft_strlen(line) - i);
-	else if (!ft_strncmp(line, "WE", i) && game->texture_pack->we->path == NULL)
-		game->texture_pack->we->path = ft_substr(line, i + 1, ft_strlen(line) - i);
-	else if (!ft_strncmp(line, "C", i) && game->texture_pack->ceiling->path == NULL)
-		game->texture_pack->ceiling->path = ft_substr(line, i + 1, ft_strlen(line) - i);
-	else if (!ft_strncmp(line, "F", i) && game->texture_pack->floor->path == NULL)
-		game->texture_pack->floor->path = ft_substr(line, i + 1, ft_strlen(line) - i);
+	if (!ft_strncmp(l, "NO", i) && game->texture_pack->no->path == NULL)
+		game->texture_pack->no->path = ft_substr(l, i + 1, ft_strlen(l) - i);
+	else if (!ft_strncmp(l, "SO", i) && game->texture_pack->so->path == NULL)
+		game->texture_pack->so->path = ft_substr(l, i + 1, ft_strlen(l) - i);
+	else if (!ft_strncmp(l, "EA", i) && game->texture_pack->ea->path == NULL)
+		game->texture_pack->ea->path = ft_substr(l, i + 1, ft_strlen(l) - i);
+	else if (!ft_strncmp(l, "WE", i) && game->texture_pack->we->path == NULL)
+		game->texture_pack->we->path = ft_substr(l, i + 1, ft_strlen(l) - i);
+	else if (!ft_strncmp(l, "C", i)
+		&& game->texture_pack->ceiling->path == NULL)
+		game->texture_pack->ceiling->path
+			= ft_substr(l, i + 1, ft_strlen(l) - i);
+	else if (!ft_strncmp(l, "F", i) && game->texture_pack->floor->path == NULL)
+		game->texture_pack->floor->path = ft_substr(l, i + 1, ft_strlen(l) - i);
 	else
 		return (0);
 	return (1);
@@ -67,6 +83,8 @@ int	game_infos(t_game *game, t_parse *control)
 		{
 			free(line);
 			line = get_next_line(control->fd);
+			if (!line)
+				return (0);
 		}
 		if (first_wall_line(line))
 			break ;
@@ -76,7 +94,9 @@ int	game_infos(t_game *game, t_parse *control)
 		line = get_next_line(control->fd);
 	}
 	if (!check_all_datas(game))
-		return (0);	// free en partant !
+		return (0); // free en partant !
 	game->map = collect_map(line, control->fd);
+	if (!game->map)
+		return (0);
 	return (1);
 }
