+++ srcs/parsing/tools.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   tools.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tpauvret <tpauvret@student.42lausanne.ch>  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/05/02 15:44:32 by tpauvret          #+#    #+#             */
+/*   Updated: 2022/05/02 15:45:19 by tpauvret         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../../inc/cub3d.h"
 
 int	is_w_space(char c)
@@ -10,12 +22,16 @@ int	is_w_space(char c)
 
 void	skip_white_space(char *str, int *i)
 {
+	if (!str)
+		return ;
 	while (is_w_space(str[*i]))
 		*i += 1;
 }
 
 void	vertical_skip_white_space(char **str, int *i, int x)
 {
+	if (!str)
+		return ;
 	while (is_w_space(str[*i][x]) && str[*i][x])
 		*i += 1;
 }
@@ -25,6 +41,8 @@ void	free_array(char **tab)
 	int	i;
 
 	i = -1;
+	if (!tab)
+		return ;
 	while (tab[++i])
 		free(tab[i]);
 	free(tab);
